[android] #3348 - Moved ExampleCustomLayer into it's own package.  Moved Custom Layer to Main Activity Controls in Nav Drawer with new icon.
